flight_controller/requirements.txt
pandas
sense_hat
sense_emu
pytest
pyserial
